requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
python-dotenv>=1.0.1
openai>=1.44.0
faiss-cpu>=1.8.0
pypdf>=4.2.0
numpy
tiktoken>=0.7.0
requests
dotenv
streamlit
